Silence json_decode warnings.
Those can happen when the string parsed is not JSON at all.
That, in turn can happen if an intermediate gateway returns it's own error,
which is ususally HTML. In any case, json_decode returns NULL when it fails
(which is what we error-check for) and warnings are not helpfull at all.